% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_loader.R
\name{pkg_loader}
\alias{pkg_loader}
\title{Package Loader and Set Up}
\usage{
pkg_loader(clear_env = F, dir = NULL, base_pkgs = T, extra_pkgs = NULL,
  sci_off = T)
}
\arguments{
\item{clear_env}{Clear global environment. Default is FALSE.}

\item{dir}{Set working directory.}

\item{base_pkgs}{Load base packages (dplyr, ggplot2, etc.). Default is TRUE.}

\item{extra_pkgs}{Load additional packages. Default is FALSE.}

\item{sci_off}{Turn off scientific notation. Default is TRUE.}
}
\description{
This function is intended to be used to load most common used packages and
other environment setups (Credit: Rob M from Staples Days)
}
\examples{
pkg_loader()
pkg_loader(clear_env = T, base_pkgs = T)
pkg_loader(base_pkgs = T, extra_pkgs = c('caret'))
}
