% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_calc.R
\name{ttest_calc}
\alias{ttest_calc}
\title{T Test Calculator}
\usage{
ttest_calc(mean1, std1, n1, mean2, std2, n2, m0 = 0, equal.variance = FALSE)
}
\arguments{
\item{mean1}{mean for sample 1}

\item{std1}{std for sample 1}

\item{n1}{size of sample 1}

\item{mean2}{mean for sample 2}

\item{std2}{std for sample 2}

\item{n2}{size of sample 2}

\item{m0}{the null value for the difference in means to be tested for. Default is 0.}

\item{equal.variance}{whether or not to assume equal variance. Default is FALSE.}

\item{df}{enter desired dataframe}

\item{vars}{variables for}
}
\value{
T=test outputs: 'Difference of means', 'Std Error', 't value', and 'p-value'.
}
\description{
This funtion is a T Test Calculator. Original script found at:
https://stats.stackexchange.com/questions/30394/how-to-perform-two-sample-t-tests-in-r-by-inputting-sample-statistics-rather-tha
}
\note{
Non equal variance is default. To change to equal variance assume, enter equal.varance = TRUE.
}
\examples{
ttest_calc(21,2,100, 25, 5, 100)
}
