% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_template.R
\name{ggplot2_template}
\alias{ggplot2_template}
\title{ggplot2 template}
\usage{
ggplot2_template()
}
\description{
This is a template function to use for ggplot2 outputs. Use this as an extension to a ggplot2 script.
}
\examples{
x_var <- factor(sample(x = 1:2, size = 100, replace = T), label = c('False', 'True'))
y_var <- sample(x = 1:100, size = 100, replace = T)
xy = data.frame(x_var, y_var)

ggplot(xy, aes(x = x_var)) + geom_bar() + ggplot2_template() + ggtitle('False vs True')
}
